refactor: separate getgoal_service from odos_goal_add_dialog #43
